+++ nextgisweb/webmap/migration/46cfda20-add-title.fw.sql
@@ -1,0 +1,7 @@
+/*** {
+    "revision": "46cfda20", "parents": ["45999710"],
+    "date": "2024-09-06T04:05:18",
+    "message": "Add title"
+} ***/
+
+ALTER TABLE webmap ADD COLUMN title character varying;