+++ nextgisweb/webmap/migration/46cfda20-add-title.rw.sql
@@ -1,0 +1,3 @@
+/*** { "revision": "46cfda20" } ***/
+
+ALTER TABLE webmap DROP COLUMN title;