+++ nextgisweb/webmap/view.py
@@ -199,8 +199,9 @@ def display(obj, request):
 
         return data
 
-    tmp = obj.to_dict()
+    title = obj.display_name if obj.title is None else obj.title
 
+    tmp = obj.to_dict()
     display_config = dict(
         extent=tmp["extent"],
         extent_const=tmp["extent_const"],
@@ -215,7 +216,7 @@ def display(obj, request):
         bookmarkLayerId=obj.bookmark_resource_id,
         webmapId=obj.id,
         webmapDescription=obj.description,
-        webmapTitle=obj.display_name,
+        webmapTitle=title,
         webmapEditable=obj.editable,
         webmapLegendVisible=obj.legend_symbols,
         drawOrderEnabled=obj.draw_order_enabled,
@@ -234,7 +235,7 @@ def display(obj, request):
             ),
         )
 
-    return dict(obj=obj, display_config=display_config, custom_layout=True)
+    return dict(obj=obj, title=title, display_config=display_config, custom_layout=True)
 
 
 @viewargs(renderer="mako")